Binding Checkboxes to GameSettings
We still need to hook up our `GameSettings.audio` setting to some user interface in the menu.

To do that we’ll start by creating a `Checkbox` widget in the usual way: a submodule in `ui.rs` and a new file in `ui/checkbox.rs`.

```rust
mod button;
mod settings;
mod checkbox;
```

The Checkbox implementation is largely not novel at this point. We’ve built a few widgets and the code inside of the Checkbox is mostly the same as the others.

The biggest difference is that we’re using the `checked` prop to control which image handle to use to display the checkbox.

<aside>
⚠️ Notably, this would normally be an `Image` widget, not a `NinePatch` because it’s a square image that we don’t intend to stretch. There’s currently a bug in the `Image` widget in kayak_ui relating to the DPI of the screen so a `NinePatch` with a border of `1.0` pixel will act the same as an image for us for now.

</aside>

```rust
use bevy::prelude::{Res, World};
use kayak_ui::{
    bevy::ImageManager,
    core::{
        rsx,
        styles::{Edge, Style, StyleProp, Units},
        widget, OnEvent, WidgetProps,
    },
    widgets::NinePatch,
};

use crate::assets::ImageAssets;

pub struct CheckboxProps {
    pub checked: bool,
    #[prop_field(OnEvent)]
    pub on_event: Option<OnEvent>,
}
pub fn Checkbox(props: CheckboxProps) {
    let styles = Style {
        width: StyleProp::Value(Units::Pixels(25.0)),
        height: StyleProp::Value(Units::Pixels(25.0)),
        ..Default::default()
    };

    let (handle_empty_box, handle_checked_box) = context
        .query_world::<Res<ImageAssets>, _, _>(|assets| {
            (
                assets.box_unchecked.clone(),
                assets.box_checked.clone(),
            )
        });

    let (empty_box, checked_box) = context
        .get_global_mut::<World>()
        .map(|mut world| {
            let mut image_manager = world
                .get_resource_mut::<ImageManager>()
                .unwrap();
            (
                image_manager.get(&handle_empty_box),
                image_manager.get(&handle_checked_box),
            )
        })
        .unwrap();

    let on_event = props.on_event.clone();

    let image = if props.checked {
        checked_box
    } else {
        empty_box
    };
    rsx! {
        <NinePatch
            on_event={on_event}
            border={Edge::all(1.0)}
            styles={Some(styles)}
            handle={image}
        />
    }
}
```

We’re going to use the `Checkbox` itself in the `SettingsMenu`. We’ll set up an `audio_checked` boolean value as well as an event handler for the checkbox, then put it inside an `Element` with some styles to make the text and checkbox display in a row.

```rust
<NinePatch
    styles={Some(container_styles)}
    border={Edge::all(10.0)}
    handle={container}
    >
    <button::SnakeButton
        on_event={Some(on_click_back)}
        >
        <Text
            size={20.0}
            content={"Back".to_string()}
        />
    </button::SnakeButton>
    <Element styles={Some(checkbox_styles)}>
        <Checkbox
            checked={audio_checked}
            on_event={Some(on_click_audio)}
        />
        <Text
            size={20.0}
            content={"Play Audio".to_string()}
        />
    </Element>
</NinePatch>
```

The `Element` styles sets a layout of `Row` with a space between the elements of `20.0` px.

```rust
let checkbox_styles = Style {
    layout_type: StyleProp::Value(LayoutType::Row),
    col_between: StyleProp::Value(Units::Pixels(20.0)),
    ..Default::default()
};
```

Before we can write more in the `SettingsMenu` we first have to head over to `ui.rs` to set up the binding to the `GameState` so that Kayak will re-render when we change it.

The system to keep the game settings up to date functions the same way the `bind_gamestate` system does. We read the `GameSettings`, check to see if they’ve changed, and update the binding.

```rust
pub fn bind_game_settings(
    state: Res<GameSettings>,
    binding: Res<Binding<GameSettings>>,
) {
    if state.is_changed() {
        binding.set(state.clone());
    }
}
```

We have to add this system to our plugin

```rust
impl Plugin for UiPlugin {
    fn build(&self, app: &mut bevy::prelude::App) {
        app.add_plugin(BevyKayakUIPlugin)
            .insert_resource(bind(STARTING_GAME_STATE))
            .add_startup_system(game_ui)
            .add_system(bind_gamestate)
            .add_system(bind_game_settings);
    }
}
```

Then in `game_ui`, we have to initialize the binding and insert it. We initialize the binding here rather than in the plugin because here we can query for the value of the `GameSettings` resource and initialize it based on that value.

Remember that we weren’t able to do that with the `GameState` because of how `iyes_loopless` inserted it, but we can here so we will.

```rust
pub fn game_ui(
    ...
    settings: Res<GameSettings>,
) {
    ...
    commands.insert_resource(bind(settings.clone()));
    ...
}
```

With the binding set up, we can use the Kayak context to query the world for the `Binding<GameSettings>` resource, and bind it to our `SettingsMenu` widget. This will cause the `SettingsMenu` to re-render when the `GameSettings` binding changes.

```rust
let settings = {
    let settings = context
        .query_world::<Res<Binding<GameSettings>>, _, _>(
            move |settings| settings.clone(),
        );

    context.bind(&settings);
    settings.get()
};
```

The checkbox click handler allows us to use the Kayak context to query the world for `GameSettings`, which we can use to swap between `ON` and `OFF` when the checkbox is clicked.

We also set up the `audio_checked` boolean that we pass to the checkbox later.

```rust
let on_click_audio = OnEvent::new(
    move |context, event| match event.event_type {
        EventType::Click(..) => {
            context.query_world::<ResMut<GameSettings>, _, _>(
        |mut settings| {
            settings.audio = match settings.audio {
                AudioSettings::ON => AudioSettings::OFF,
                AudioSettings::OFF => AudioSettings::ON,
            };
        },
    );
        }
        _ => {}
    },
);

let audio_checked = match settings.audio {
    AudioSettings::ON => true,
    AudioSettings::OFF => false,
};
```

Now when we `cargo run` the audio can be controlled through the settings pane by the player.

An unchecked “Play Audio” box correctly indicates that no sounds are played, while a checked box indicates that sounds do play.