building-a-snake-resources-components-and-bundles
In the previous lesson we talked about storing the index of the grid position in the Snake's VecDeque. We're going to have more interesting logic in general so instead of storing `(x,y)` indexes in the Snake, we're going to store `Entity` ids.

Each of those `Entity` ids is going to refer to a separate segment of the snake that we've spawned into the game. This will allow us to keep a fairly simple model of our Snake and the order of the segments, which in turn will allow us to identify both the head, tail, and surrounding segments when we want to render different sprites for the snake segments.

We have to do a few things here.

* Create a `Component` named `Position`
* Create a `Resource` named `Snake`
* Get access to the `TILE_SIZE`
* Get access to the `cell_position_to_physical` function
* Spawn in the snake segments

\## Creating a Component named Position

We want a way to refer to cells in our 2d board grid. We'll query for this `Component` in the systems that control spawning elements like snake segments or apples as well as when we're updating the rendering the snake body sprites.

We could make our own component like this and that would work just fine.

```rust
\#[derive(Component)]
struct Position {
	x: i8,
	y: i8
}
```

but Bevy already has a struct that can hold an `x` value and a `y` value which are both integers: [`IVec2`](https://docs.rs/bevy/0.12.0/bevy/prelude/struct.IVec2.html).

We can't insert an `IVec2` on an `Entity` as a `Component` though, primarily because it doesn't implement the `Component` trait, but also because then we'd have trouble differentiating it from any other `IVec2` that could be inserted in the future.

Instead we can take advantage of a pattern called the "newtype" pattern. The newtype pattern is a really lightweight pattern. It is literally wrapping some type in our own type.

In our case, we'll create a tuple struct that wraps the `IVec2`. We'll call it `Position`.

This allows us to differentiate `IVec2`s that we're using to represent `Position`s on the board from any other `IVec2` using the Rust compiler.

Place this code in the `src/board.rs` file, since it refers to a board position.

```rust
\#[derive(Debug, Component)]
pub struct Position(pub IVec2);
```

\## Create a Resource named Snake

We're going to build a `Resource` called `Snake` that will hold the `Entity` ids for the snake segments we'll be spawning in.

In `src/lib.rs` create a new module called `snake`.

```rust
pub mod board;
pub mod colors;
pub mod snake;
```

Inside `src/snake.rs` we're going to define our `Snake`. This `Snake` will derive `Resource` so that we can use it as a Bevy `Resource`, `Debug` so we can `dbg!` it out, and `Default` so that we can initialize an empty `Snake`.

The derive'd `Default` implementation will create an empty `VecDeque` for us, so we don't have to worry about what a "default" `Entity` would be, because we won't have any in the default case.

```rust
use bevy::prelude::*;
use std::collections::VecDeque;

\#[derive(Debug, Default, Resource)]
pub struct Snake {
    pub segments: VecDeque<Entity>,
}
```

We can insert our `Snake` into the game in `src/main.rs`. Note that we're using [`init_resource`](https://docs.rs/bevy/0.12.0/bevy/app/struct.App.html#method.init_resource) instead of [`insert_resource`](https://docs.rs/bevy/0.12.0/bevy/app/struct.App.html#method.insert_resource) here.

`init_resource` will use our `Default` implementation to initialize a new `Snake`, while `insert_resource` allows us to specify value to use.

```rust
.insert_resource(Board::new(20))
.init_resource::<Snake>()
.add_systems(Startup, (setup, spawn_board))
.run();
```

`init_resource` takes a type argument to determine what type is being initialized. In this case, we're specifying `Snake` using syntax that's affectionately called "turbofish syntax" because [it kind of looks like a fish](https://turbo.fish/): `::<>`.

\## Spawning Snake Segments

We can expand on the code in `snake.rs` with a system called `spawn_snake` that will handle spawning in a snake's segments and populating the `Snake`'s `Entity`s.

`spawn_snake` needs access to `commands` to spawn in `Sprite`s, our `Board` as a read-only `Resource` ([`Res`](https://docs.rs/bevy/0.12.0/bevy/prelude/struct.Res.html)), and `Snake` as a mutable `Resource` ([`ResMut`](https://docs.rs/bevy/0.12.0/bevy/prelude/struct.ResMut.html)).

```rust
use bevy::prelude::*;
use std::collections::VecDeque;

use crate::{
    board::{Board, Position},
    colors,
};

\#[derive(Debug, Default, Resource)]
pub struct Snake {
    pub segments: VecDeque<Entity>,
}

pub fn spawn_snake(
    mut commands: Commands,
    board: Res<Board>,
    mut snake: ResMut<Snake>,
) {
    for position in [
        Position(IVec2::new(3, 4)),
        Position(IVec2::new(4, 4)),
    ] {
        let entity = commands
            .spawn((
                SpriteBundle {
                    sprite: Sprite {
                        color: colors::SNAKE,
                        custom_size: Some(Vec2::splat(
                            TILE_SIZE,
                        )),
                        ..default()
                    },
                    transform: Transform::from_xyz(
                        board.cell_position_to_physical(
                            position.0.x,
                        ),
                        board.cell_position_to_physical(
                            position.0.y,
                        ),
                        2.0,
                    ),
                    ..default()
                },
                position,
            ))
            .id();

        snake.segments.push_front(entity);
    }
}
```

`spawn_snake` iterates over two `Position`s and `spawn`s in a `SpriteBundle`. For each `Entity` we create, we push that `Entity` id into the `snake.segments`.

To get this to work though, we're going to need access to `TILE_SIZE` and `cell_position_to_physical`.

We have some options for both.

For `TILE_SIZE`, which is a const in the `board.rs` file, we can add a `pub` label and bring it into scope, or we can add a function that will let us access the `TILE_SIZE` through the `Board`.

Either way works. I'll add `pub` to the `const` in `board.rs`, but I encourage you to implement this as a function on `Board` that returns a `Vec2` if you want some extra practice.

```rust
pub const TILE_SIZE: f32 = 30.0;
```

Then bring `TILE_SIZE` into scope in `snake.rs`.

```rust
use crate::{
    board::{Board, Position, TILE_SIZE},
    colors,
};
```

For `cell_position_to_physical`, we'll do the same. Add `pub` in `board.rs` and we'll be able to use it wherever we have a `Board`.

```rust
pub fn cell_position_to_physical(
    &self,
    pos: i32,
) -> f32 {
    let offset =
        -self.physical_size / 2.0 + 0.5 * TILE_SIZE;

    offset
        + pos as f32 * TILE_SIZE
        + (pos + 1) as f32 * TILE_SPACER
}
```

\## Snake Color

Then we need to add a color to our `colors.rs` file. I'll use white for now, but later we'll be using actual snake textures so this color is temporary and you can use whatever you want.

```rust
pub const SNAKE: Color = Color::WHITE;
```

\## Adding the snake

In `src/main.rs`, we still need to actually spawn the snake in. Bring the new `spawn_snake` function into scope.

```rust
use snake::{
    board::{spawn_board, Board},
    snake::{spawn_snake, Snake},
};
```

and then add `spawn_snake` to the end of our `Startup` systems.

```rust
.add_systems(
    Startup,
    (setup, spawn_board, spawn_snake),
)
```

Now when using `cargo run`, you should see two squares rendered as the snake color you chose.

![[first-snake-appearance.png]]