moving-the-snake
Our snake is on the board, and we’ve got a system executing on each tick. It’s time to make the snake move. We'll do this by spawning a new segment on the next tile the snake is going to move to and then despawning its tail. This will keep the snake the same length as it moves across the board.

In `lib.rs`, bring `bevy::prelude::*`, `Position`, and `Snake` into scope.

```rust
use bevy::prelude::*;
use board::Position;
use snake::Snake;
```

Then in `tick` we'll add 3 arguments. `commands`, `snake`, and `positions`.

We'll use `commands` for spawning entities and `snake` for access to the snake segments.

With `positions`, we're querying for all of the entities with a `Position` component. We can use this effectively like a database later, doing lookups for a specific `Position` component related to a specific `Entity`.

```rust
pub fn tick(
    mut commands: Commands,
    mut snake: ResMut<Snake>,
    positions: Query<&Position>,
) {
    let snake_head_entity = snake
        .segments
        .front()
        .expect("snake should have a head entity");

	let next_position = positions.get(*snake_head_entity)
	    .map(|head| {
	        Position::new(head.x + 1, head.y)
	    })
	    .expect("stored entities in a snake should have a Position component associated with them");

    // TODO: Spawn sprite in here
    let entity = commands.spawn(next_position).id();

    snake.segments.push_front(entity);

    let old_tail = snake
        .segments
        .pop_back()
        .expect("snake should have a tail entity");
    commands.entity(old_tail).despawn_recursive();

    info!(?snake);
}
```

```rust
let snake_head_entity = snake
    .segments
    .front()
    .expect("snake should have a head entity");
```

Using a [`VecDeque`](https://doc.rust-lang.org/std/collections/struct.VecDeque.html#method.front) starts to pay off for us here. `snake.segments` is our `VecDeque` which has a function called [`front`](https://doc.rust-lang.org/std/collections/struct.VecDeque.html#method.front) associated with it. `front` returns an `Option<&T>`, which is an `Option` that contains a shared reference to the type of the items in our snake segments. In this case, that's `Option<&Entity>`.

We'll use [`.expect`](https://doc.rust-lang.org/std/option/enum.Option.html#method.expect) to unwrap the `Option` because our game doesn't work if the `Snake` doesn't have a head segment, so `panic`ing here is ok.

That leaves us with `&Entity`, which is a shared reference to an [`Entity`](https://docs.rs/bevy/0.12.0/bevy/ecs/entity/struct.Entity.html).

```rust
let next_position = positions.get(*snake_head_entity)
    .map(|head| {
        Position::new(head.x + 1, head.y)
    })
    .expect("stored entities in a snake should have a Position component associated with them");
```

When we query for components with [`Query`](https://docs.rs/bevy/0.12.0/bevy/prelude/struct.Query.html), we can use that query result as if it was a database. For us, that means we can use `positions` as a database that lets us access the `Position` for a particular `Entity`.

To do this, we need to use [`.get`](https://docs.rs/bevy/0.12.0/bevy/prelude/struct.Query.html#method.get) with an argument of `Entity`. Since `snake_head_entity` is a shard reference: `&Entity`, we need to de-reference the shared reference with `*`.

This gives us a `Result<&Position, QueryEntityError>`.

We're going to use this `&Position` to create the next position for our snake to be in. The `Result` type has a function called [`.map`](https://doc.rust-lang.org/std/result/enum.Result.html#method.map) that allows us to operate on the inner value of `Ok`, if it exists. If the `Result` is an `Err` value, then `.map` will not run.

We'll take advantage of `.map` and use the head position of our snake to place our new position to the right of the current snake head. `.map` allows us to return any value we want so we'll return a new `Position` with the new values.

Finally, we have to deal with the `Result`. In this case, our game won't work correctly if we didn't get a result from `positions.get`, so we can use `.expect` to unwrap the `Result` if we have an `Ok` value, and panic if we have an `Err` value.

> [!NOTE] .expect message style
> There is a style guide for writing `.expect` messages available in the [`Result.expect` docs](https://doc.rust-lang.org/std/result/enum.Result.html#recommended-message-style).

```rust
// TODO: Spawn sprite in here
let entity = commands.spawn(next_position).id();
```

[`commands.spawn`](https://docs.rs/bevy/0.12.0/bevy/ecs/system/struct.Commands.html#method.spawn) will create an `Entity` for us and allows us to insert any components we want to associate with that `Entity`. In this case, we're inserting the `Position` component we created in `next_position`. Then we return the `Entity` using the `id()` function.

```rust
snake.segments.push_front(entity);

let old_tail = snake
    .segments
    .pop_back()
    .expect("snake should have a tail entity");

commands.entity(old_tail).despawn_recursive();
```

The `VecDeque` comes in handy once again, allowing us to push the `Entity` we just spawned onto the front of the `Snake` using [`push_front`](https://doc.rust-lang.org/std/collections/struct.VecDeque.html#method.push_front) so we can do it all again later.

Then we can take advantage of [`pop_back`](https://doc.rust-lang.org/std/collections/struct.VecDeque.html#method.pop_back) to pop the `Entity` that represents the tail of the `Snake` off of the `VecDeque`. `pop_back` returns an `Option<Entity>` because the `VecDeque` could be empty (and thus not have a value to pop off), but if the `VecDeque` is empty that means we have some major problems in our game, so we'll choose to use `.expect` once again to unwrap the `Entity` or panic if it doesn't exist.

Since we've moved the Snake forward one square already that means we have three squares of snake segments currently. We need to despawn the tail `Entity` and all components associated with it.

To do that, we use [`commands.entity`](https://docs.rs/bevy/0.12.0/bevy/ecs/system/struct.Commands.html#method.entity) to get access to the [`EntityCommands`](https://docs.rs/bevy/0.12.0/bevy/ecs/system/struct.EntityCommands.html) which has an implementation that gives us access to [`despawn_recursive`] (https://docs.rs/bevy/0.12.0/bevy/hierarchy/trait.DespawnRecursiveExt.html#tymethod.despawn_recursive) which de-spawns the `Entity` and all of its descendants.

```rust
info!(?snake);
```

Finally, we use Bevy's tracing support to log out the `Debug` representation of our `Snake`. The `?` prefix on our `snake` variable is called a sigil and indicates to the tracing crate that the `Debug` representation should be used.

After running the game with `cargo run`, note that each snake receives a new `Entity` id as the system runs each time.

```
2023-11-19T23:12:42.987930Z  INFO snake: snake=ResMut(Snake { segments: [405v9, 3v9] })
2023-11-19T23:12:43.088440Z  INFO snake: snake=ResMut(Snake { segments: [2v10, 405v9] })
2023-11-19T23:12:43.188419Z  INFO snake: snake=ResMut(Snake { segments: [3v10, 2v10] })
```